// generated/src/app/api-models/auth-procedure.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuthProcedure {
  authNumber: number;
  secondaryAuthNo: string;
  seqAuthDetail: number;
  fromDate: Date;
  thruDate: Date;
  authProcedures: string;
  authorizedQty: number;
  allowedAmt: number;
  status: string;
  statusDate: Date;
  statusReason: string;
  denialNotification: string;
  notificationDate: Date;
  actualQty: number;
  actualAmt: number;
  medDefCode: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  costProcedureCd: string;
  priceSchedule: string;
  priceRegion: string;
  authorizedCost: number;
  fromDate1: Date;
  thruDate1: Date;
  authProcedure1: string;
  authorizedQty1: number;
  actualQty1: number;
  actualAmt1: number;
  allowedAmt1: number;
  costProcedureCd1: string;
  priceSchedule1: string;
  priceRegion1: string;
  authorizedCost1: number;
  fromDate2: Date;
  thruDate2: Date;
  authProcedure2: string;
  authorizedQty2: number;
  actualQty2: number;
  actualAmt2: number;
  allowedAmt2: number;
  costProcedureCd2: string;
  priceSchedule2: string;
  priceRegion2: string;
  authorizedCost2: number;
  fromDate3: Date;
  thruDate3: Date;
  authProcedure3: string;
  authorizedQty3: number;
  actualQty3: number;
  actualAmt3: number;
  allowedAmt3: number;
  costProcedureCd3: string;
  priceSchedule3: string;
  priceRegion3: string;
  authorizedCost3: number;
  requestedAmount: number;
  requestedQty: number;
  purchasePrice: number;
  ndcCode: string;
  toothNumber: string;
  surface1: string;
  surface2: string;
  surface3: string;
  surface4: string;
  surface5: string;
  arch: string;
  quadrant: string;
  oralCavity: string;
  paperworkAttached: string;
  providerTrackingNo: string;
  providerEntityId: string;
  clrnghseTrackingNo: string;
  clrnghseEntityId: string;
  responseTrackingNo: string;
  responseEntityId: string;
  processStatus: string;
}